requirements.txt
discord.py==1.2.5
emoji==0.5.3